+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## DEV
 
+ * Do not show pretty errors if "text/html" is not accepted (see #249).
  * Return `null` in `Model::findMultipleByIds()` if there are no models (see #266).
  * Restore compatibility with Doctrine DBAL 2.9 (see #256).